fill_in_soil_gaps now does all files at once, should be faster

+++ util/fill_in_soil_gaps.sh
@@ -2,10 +2,11 @@
 
 ## Parameters
 proj=hd50
-year=2010
-month=07
-user=jt4085
-ens_to_check="003"
+year=2018
+month=02
+user=*
+ens_to_check="000"
+ens_to_check="000 001 002 003 004 005 006 007 008 009 ctl"
 steams_to_check="SLV1H"
 vars_to_check="soil_mois soil_mois_frozen_frac"
 
@@ -40,6 +41,8 @@ for cycle_dir in $root_dir/*; do
                 exit 1
             fi
 
+            # Construct a list of files to extract
+            declare -a action_list=()
             for var in $vars_to_check; do
                 echo -e "\t\t\t$var"
 
@@ -49,35 +52,51 @@ for cycle_dir in $root_dir/*; do
                 else
                     echo -e "\t\t\t\tFile not found in tarball"
 
-                    echo -e "\t\t\t\tFetching tarball from MDSS"
-                    mdss -P $proj get barra2/barra-r2/prod/$user/cg406*/$year/$month/$cycle/$ens/nc/$stream.tar.gz $temp_file_dir
-
-                    echo -e "\t\t\t\tUnpacking file from tarball"
-                    new_tarball=$temp_file_dir/${stream}.tar.gz
-
-                    # Create the directory structure for the old tar
-                    mkdir -p $temp_file_dir/nc/$stream
-
-                    # Get the name of the file to extract
-                    file_to_extract=`tar -tvf $new_tarball | \
-                                     grep $regex | \
-                                     awk '{print $NF}'`
-
-                    # Extract the file
-                    tar -zx --directory=$temp_file_dir/nc/$stream \
-                        --file=$new_tarball $file_to_extract
-
-
-                    echo -e "\t\t\t\tAdding file to old tarball"
-                    tar --append --file=$tarball_path  --directory=$temp_file_dir nc/$stream/$(basename $file_to_extract)
-
-                    # Remove the tarball just in case
-                    rm $new_tarball
-
+                    action_list+=("${regex}*.nc")
                 fi
             done
-        done
 
+            echo
+
+            if [[ ${#action_list[@]} > 0 ]]; then
+                echo -e "\t\t\tFetching tarball from MDSS"
+                mdss_path=barra2/barra-r2/prod/$user/cg406*/$year/$month/$cycle/$ens/nc/$stream.tar.gz
+                mdss -P $proj get $mdss_path $temp_file_dir
+
+                echo -e "\t\t\tUnpacking files from tarball"
+                new_tarball=$temp_file_dir/${stream}.tar.gz
+
+                # Clear out the extraction dest first
+                # Any existing files will make a mess with wildcards below
+                rm -rf $temp_file_dir/nc/$stream
+
+                # Create the directory structure for the old tar
+                mkdir -p $temp_file_dir/nc/$stream
+
+                # Extract the files
+                tar -zx \
+                    --file=$new_tarball \
+                    --directory=$temp_file_dir/nc/$stream \
+                    --wildcards \
+                    ${action_list[*]}
+
+                # Add nc/$stream/ infront of each file name
+                # Use a subshell and cd to handle wildcards
+                echo -e "\t\t\tAdding files to old tarball"
+                (
+                 cd $temp_file_dir &&
+                 tar --append \
+                     --file=$tarball_path \
+                     --wildcards \
+                     ${action_list[*]/#/nc/$stream/}
+                )
+
+                # Remove the tarball just in case
+                rm $new_tarball
+            else
+                echo -e "\t\t\tNothing to do"
+            fi
+        done
     done
 done
